+++ tomcat/src/main/java/nextstep/jwp/controller/other/OtherController.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.controller.other;
 
-import static org.apache.coyote.http11.common.FileContent.STATIC;
+import static nextstep.jwp.controller.FileContent.STATIC;
 
 import java.io.File;
 import java.io.IOException;
